Move HashFunction type into separate file in @iov/crypto
