+++ packages/iov-crypto/src/hmac.ts
@@ -1,4 +1,4 @@
-import { HashFunction } from "./sha";
+import { HashFunction } from "./hash";
 
 export class Hmac<H extends HashFunction> implements HashFunction {
   public readonly blockSize: number;
