+++ packages/iov-crypto/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./bip39";
 export { EnglishMnemonic } from "./englishmnemonic";
+export * from "./hash";
 export * from "./hmac";
 export * from "./keccak";
 export * from "./libsodium";
