+++ packages/iov-crypto/src/sha.ts
@@ -1,10 +1,6 @@
 import shajs from "sha.js";
 
-export interface HashFunction {
-  readonly blockSize: number;
-  readonly update: (_: Uint8Array) => HashFunction;
-  readonly digest: () => Uint8Array;
-}
+import { HashFunction } from "./hash";
 
 export class Sha1 implements HashFunction {
   public readonly blockSize = 512 / 8;
