+++ packages/iov-crypto/types/hash.d.ts
@@ -1,0 +1,5 @@
+export interface HashFunction {
+  readonly blockSize: number;
+  readonly update: (_: Uint8Array) => HashFunction;
+  readonly digest: () => Uint8Array;
+}
